fix: do not store incomplete config from increase_get_serial()
`increase_get_serial()` is called when `import` CLI command is used.
If the bot is not configured, previously it stored an incomplete
config with only the `serial` field. Such bot config cannot
be deserialized into `BotConfig` structure afterwards.

The function is changed not to store the serial and always return 1
until the bot is started at least once and users get a chance
to receive an update.